Update
Opraveny přístupy pro nepřihlášené uživatele
Vyřešen problém se špatným ukladánáním údajů pro knihy
Odstranění možnosti smazat autora knihy v případě, že je k nějaké přiřazen
Opraveno ukládání fotografií a editace záznamů

+++ backend/controllers/BorrowedbooksController.php
@@ -4,9 +4,13 @@ namespace backend\controllers;
 
 use frontend\models\Borrowedbooks;
 use frontend\models\BorrowedbooksSearch;
+use frontend\models\Storedbooks;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\base\Exception;
+use yii\base\Model;
+use Yii;
 
 /**
  * BorrowedbooksController implements the CRUD actions for Borrowedbooks model.
@@ -38,13 +42,19 @@ class BorrowedbooksController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new BorrowedbooksSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
-
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+        if(Yii::$app->user->identity != NULL) {
+            $searchModel = new BorrowedbooksSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
+    
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -55,9 +65,15 @@ class BorrowedbooksController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(Yii::$app->user->identity != NULL) {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -67,19 +83,25 @@ class BorrowedbooksController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Borrowedbooks();
-
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                return $this->redirect(['view', 'id' => $model->id]);
+        if(Yii::$app->user->identity != NULL) {
+            $model = new Borrowedbooks();
+    
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post()) && $model->save()) {
+                    return $this->redirect(['view', 'id' => $model->id]);
+                }
+            } else {
+                $model->loadDefaultValues();
             }
-        } else {
-            $model->loadDefaultValues();
+    
+            return $this->render('create', [
+                'model' => $model,
+            ]);
         }
-
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -91,15 +113,21 @@ class BorrowedbooksController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
-
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+        if(Yii::$app->user->identity != NULL) {
+            $model = $this->findModel($id);
+    
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+    
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -111,9 +139,17 @@ class BorrowedbooksController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if(Yii::$app->user->identity != NULL) {
+            $model = $this->findModel($id);
+    
+            $this->findModel($id)->delete();
+    
+            return $this->redirect(['index']);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
