+++ backend/controllers/StoredbooksController.php
@@ -9,6 +9,9 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\web\UploadedFile;
+use yii\base\Exception;
+use yii\base\Model;
+use Yii;
 
 /**
  * StoredbooksController implements the CRUD actions for Storedbooks model.
@@ -40,13 +43,19 @@ class StoredbooksController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new StoredbooksSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
-
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+        if(Yii::$app->user->identity != NULL) {
+            $searchModel = new StoredbooksSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
+    
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -57,9 +66,20 @@ class StoredbooksController extends Controller
      */
     public function actionView($id)
     {
+        /*
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
+        */
+        if(Yii::$app->user->identity != NULL) {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -69,48 +89,55 @@ class StoredbooksController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Storedbooks();
-        $modela = new Authors();
-
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $modela->load($this->request->post()) ) {
-                if($_POST['Authors']['id']==NULL)
-                {
-                    return $this->redirect(['create']);
-                }
-                else
-                {
-                    $model->authorid = $_POST['Authors']['id'];
-                    $model->save(false);
-                    $modela->save();
-
-                    //*******************************************************************************
-
-                    $bookid=$model->id;
-                    $image = UploadedFile::getInstance($model, 'img');
-                    if($model->img!="") {
-                        $imageName = 'book' . $bookid . '.' . $image->getExtension();
-                        $image->saveAs('D:/xampp/htdocs/knihovnakucera/images/books/' . $imageName);
-                        $model->img = $imageName;
+        if(Yii::$app->user->identity != NULL) {
+            $model = new Storedbooks();
+            $modela = new Authors();
+    
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post()) && $modela->load($this->request->post()) ) {
+                    if($_POST['Authors']['id']==NULL)
+                    {
+                        return $this->redirect(['create']);
                     }
-                    else{
-                        $model->img = "default.png";
+                    else
+                    {
+                        $model->authorid = $_POST['Authors']['id'];
+                        $model->save(false);
+                        $modela->save();
+                        //*******************************************************************************
+        
+                        $bookid = $model->id;
+                        $image = UploadedFile::getInstance($model, 'img');
+        
+                        if($image) {   //pokud je fileinput vyplnen, tak probehne. V pripade, ze ne, tak se provede else a nastavi se defaultni foto
+                            $imageName = 'book' . $bookid . '.' . $image->getExtension();
+                            $image->saveAs('@frontend/images/books/' . $imageName);
+                            $model->img = $imageName;
+                        }
+                        else{
+                            $model->img = "default.png";
+                        }
+        
+                        $model->save();
+        
+                        return $this->redirect(['view', 'id' => $model->id]);
                     }
-                    $model->save();
-
-                    return $this->redirect(['view', 'id' => $model->id]);
                 }
             }
+            else {
+                $model->loadDefaultValues();
+                $modela->loadDefaultValues();
+            }
+    
+            return $this->render('create', [
+                'model' => $model,
+                'modela' => $modela,
+            ]);
         }
-        else {
-            $model->loadDefaultValues();
-            $modela->loadDefaultValues();
+        else
+        {
+            return $this->render('/site/index');
         }
-
-        return $this->render('create', [
-            'model' => $model,
-            'modela' => $modela,
-        ]);
     }
 
     /**
@@ -122,43 +149,49 @@ class StoredbooksController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
-        $modela = new Authors();
-
-        if ($this->request->isPost && $model->load($this->request->post()) && $modela->load($this->request->post())) {
-            if($_POST['Authors']['id']==NULL)   //pokud je pri updatu nevyplneny autor, tak update neprobehne.
-            {
-                return $this->redirect(['update','id' => $model->id]);
-            }
-            else {
-                $model->authorid = $_POST['Authors']['id'];
-                $model->save(false);
-                $modela->save();
-
-                //*******************************************************************************
-
-                $bookid = $model->id;
-                $image = UploadedFile::getInstance($model, 'img');
-
-                if($image) {   //pokud je fileinput vyplnen, tak probehne. V pripade, ze ne, tak se provede else a nastavi se defaultni foto
-                    $imageName = 'book' . $bookid . '.' . $image->getExtension();
-                    $image->saveAs('D:/xampp/htdocs/knihovnakucera/images/books/' . $imageName);
-                    $model->img = $imageName;
+        if(Yii::$app->user->identity != NULL) {
+            $model = $this->findModel($id);
+            $modela = new Authors();
+    
+            if ($this->request->isPost && $model->load($this->request->post()) && $modela->load($this->request->post())) {
+                if($_POST['Authors']['id']==NULL)   //pokud je pri updatu nevyplneny autor, tak update neprobehne.
+                {
+                    return $this->redirect(['update','id' => $model->id]);
                 }
-                else{
-                    $model->img = "default.png";
+                else {
+                    $model->authorid = $_POST['Authors']['id'];
+                    $model->save(false);
+                    $modela->save();
+    
+                    //*******************************************************************************
+    
+                    $bookid = $model->id;
+                    $image = UploadedFile::getInstance($model, 'img');
+    
+                    if($image) {   //pokud je fileinput vyplnen, tak probehne. V pripade, ze ne, tak se provede else a nastavi se defaultni foto
+                        $imageName = 'book' . $bookid . '.' . $image->getExtension();
+                        $image->saveAs('@frontend/images/books/' . $imageName);
+                        $model->img = $imageName;
+                    }
+                    else if($model->img == "" || $model->img == "default.png"){
+                        $model->img = "default.png";
+                    }
+    
+                    $model->save();
+    
+                    return $this->redirect(['view', 'id' => $model->id]);
                 }
-
-                $model->save();
-
-                return $this->redirect(['view', 'id' => $model->id]);
             }
+    
+            return $this->render('update', [
+                'model' => $model,
+                'modela' => $modela,
+            ]);
         }
-
-        return $this->render('update', [
-            'model' => $model,
-            'modela' => $modela,
-        ]);
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -170,13 +203,19 @@ class StoredbooksController extends Controller
      */
     public function actionDelete($id)
     {
-        $data = Storedbooks::findOne($id);
-        if($data->img != "default.png")
-            unlink('D:/xampp/htdocs/knihovnakucera/images/books/' . $data->img);
-
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if(Yii::$app->user->identity != NULL) {
+            $data = Storedbooks::findOne($id);
+            if($data->img != "default.png")
+                unlink(\Yii::getAlias('@frontend').'/images/books/' . $data->img);
+    
+            $this->findModel($id)->delete();
+    
+            return $this->redirect(['index']);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -193,5 +232,102 @@ class StoredbooksController extends Controller
         }
 
         throw new NotFoundHttpException('The requested page does not exist.');
+    }
+    /**
+     * Odstrani foto u knihy pokud je nejaka nastavena
+     * @param int $id ID
+     * @return \yii\web\Response
+     */
+    public function actionOdstranitfoto($id)
+    {
+        if(Yii::$app->user->identity != NULL) {
+            $model=Storedbooks::findOne(['id' => $id]);
+            
+            if($model->img != "default.png")
+                unlink(\Yii::getAlias('@frontend').'/images/books/' . $model->img);
+                
+            $model->img="default.png";
+            $model->save();
+    
+            return $this->redirect(['update', 'id' => $model->id]);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
+    }
+    /**
+     * Zapise ke knize v databázi hodnotu 0, která představuje vracenou knihu
+     * smaze zaznam z tabulky borrowedbooks (neni uz potreba)
+     * @param int $id ID
+     * @return \yii\web\Response
+     */
+    public function actionVratit($id)
+    {
+        if(Yii::$app->user->identity != NULL) {
+            $model=Storedbooks::findOne(['id' => $id]);
+            Borrowedbooks::deleteAll(['idbook'=>$model->id]);
+            $model->save();
+    
+            return $this->redirect(['index']);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
+    }
+    /**
+     * Zapise ke knize v databázi hodnotu 1, která představuje vypůjčení knihy, pricte pocet pujceni,
+     * vytvori novy zaznam v databazi borrowedbooks, ktery uchova po dobu pujceni potrebna data
+     * @param int $id ID
+     * @return \yii\web\Response
+     */
+    public function actionPujcit($id)
+    {
+        /*
+        $model=Storedbooks::findOne(['id' => $id]);
+        $modela=new Borrowedbooks();
+        $datum=date("Y-m-d");
+        $mod_date = strtotime($datum."+ 14 days");
+
+        $modela->idbook=$model->id;
+        $modela->iduser=Yii::$app->user->getId();
+        $modela->fromdate=date($datum);
+        $modela->untildate=date("Y-m-d",$mod_date);
+        $modela->save();
+        
+        $model->borrowedcount++;
+        $model->save();
+        */
+        if(Yii::$app->user->identity != NULL) {
+            $transaction = Yii::$app->db->beginTransaction();
+            try {
+                $model=Storedbooks::findOne(['id' => $id]);
+                $modela=new Borrowedbooks();
+                $datum=date("Y-m-d");
+                $mod_date = strtotime($datum."+ 14 days");
+    
+                $modela->idbook=$model->id;
+                $modela->iduser=Yii::$app->user->getId();
+                $modela->fromdate=date($datum);
+                $modela->untildate=date("Y-m-d",$mod_date);
+    
+                $model->borrowedcount++;
+    
+                $model->save();
+                $modela->save();
+    
+                $transaction->commit();
+            } catch (\Exception $e) {
+                $transaction->rollBack();
+                throw $e;
+            }
+    
+            return $this->redirect(['index']);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 }
