+++ frontend/controllers/AuthorsController.php
@@ -8,6 +8,9 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\web\UploadedFile;
+use yii\base\Exception;
+use yii\base\Model;
+use Yii;
 
 /**
  * AuthorsController implements the CRUD actions for Authors model.
@@ -39,13 +42,19 @@ class AuthorsController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new AuthorsSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
-
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+        if(Yii::$app->user->identity != NULL) {
+            $searchModel = new AuthorsSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
+    
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -56,9 +65,15 @@ class AuthorsController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(Yii::$app->user->identity != NULL) {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -68,31 +83,37 @@ class AuthorsController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Authors();
-
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post())) {
-
-                $image = UploadedFile::getInstance($model, 'img');
-                if($image) {
-                    $imageName = 'author' . $model->id . '.' . $image->getExtension();
-                    $image->saveAs('D:/xampp/htdocs/knihovnakucera/images/authors/' . $imageName);
-                    $model->img = $imageName;
+        if(Yii::$app->user->identity != NULL) {
+            $model = new Authors();
+    
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post())) {
+    
+                    $image = UploadedFile::getInstance($model, 'img');
+                    if($image) {
+                        $imageName = 'author' . $model->id . '.' . $image->getExtension();
+                        $image->saveAs('D:/xampp/htdocs/knihovnakucera/images/authors/' . $imageName);
+                        $model->img = $imageName;
+                    }
+                    else{
+                        $model->img = "default.png";
+                    }
+                    $model->save();
+    
+                    return $this->redirect(['view', 'id' => $model->id]);
                 }
-                else{
-                    $model->img = "default.png";
-                }
-                $model->save();
-
-                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                $model->loadDefaultValues();
             }
-        } else {
-            $model->loadDefaultValues();
+    
+            return $this->render('create', [
+                'model' => $model,
+            ]);
         }
-
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -104,26 +125,32 @@ class AuthorsController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
-
-        if ($this->request->isPost && $model->load($this->request->post())) {
-            $image = UploadedFile::getInstance($model, 'img');
-            if($image) {
-                $imageName = 'author' . $model->id . '.' . $image->getExtension();
-                $image->saveAs('D:/xampp/htdocs/knihovnakucera/images/authors/' . $imageName);
-                $model->img = $imageName;
+        if(Yii::$app->user->identity != NULL) {
+            $model = $this->findModel($id);
+    
+            if ($this->request->isPost && $model->load($this->request->post())) {
+                $image = UploadedFile::getInstance($model, 'img');
+                if($image) {
+                    $imageName = 'author' . $model->id . '.' . $image->getExtension();
+                    $image->saveAs('D:/xampp/htdocs/knihovnakucera/images/authors/' . $imageName);
+                    $model->img = $imageName;
+                }
+                else{
+                    $model->img = "default.png";
+                }
+                $model->save();
+    
+                return $this->redirect(['view', 'id' => $model->id]);
             }
-            else{
-                $model->img = "default.png";
-            }
-            $model->save();
-
-            return $this->redirect(['view', 'id' => $model->id]);
+    
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
@@ -135,9 +162,15 @@ class AuthorsController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if(Yii::$app->user->identity != NULL) {
+            $this->findModel($id)->delete();
+    
+            return $this->redirect(['index']);
+        }
+        else
+        {
+            return $this->render('/site/index');
+        }
     }
 
     /**
